# docs/openapi.yaml
openapi: 3.1.0
info:
  title: WAP — WordPress AI Platform
  version: 0.1.0
paths:
...
  /admin/agents:
    get:
      summary: List Agents
      description: List all agent definitions (mapped and unmapped).
      operationId: list_agents_admin_agents_get
      security:
      - HTTPBearer: []
      parameters:
      - name: admin_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Admin Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response List Agents Admin Agents Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      summary: Create Agent
      description: 'Create a new agent definition.


        The agent is not mapped to any role — it is an undeployed draft.

        Use PUT /roles/{role} to assign it.'
      operationId: create_agent_admin_agents_post
      security:
      - HTTPBearer: []
      parameters:
      - name: admin_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Admin Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Create Agent Admin Agents Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/agents/{agent_id}:
    get:
      summary: Get Agent
      operationId: get_agent_admin_agents__agent_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: agent_id
        in: path
        required: true
        schema:
          type: string
          title: Agent Id
      - name: admin_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Admin Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Agent Admin Agents  Agent Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      summary: Update Agent
      description: Update an agent definition.  Triggers registry reload for this
        agent.
      operationId: update_agent_admin_agents__agent_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: agent_id
        in: path
        required: true
        schema:
          type: string
          title: Agent Id
      - name: admin_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Admin Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Update Agent Admin Agents  Agent Id  Put
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      summary: Delete Agent
      description: 'Hard-delete an agent.


        Fails with 409 if the agent is currently mapped to any role.'
      operationId: delete_agent_admin_agents__agent_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: agent_id
        in: path
        required: true
        schema:
          type: string
          title: Agent Id
      - name: admin_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Admin Session
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/roles:
    get:
      summary: List Roles
      description: List all role → agent_id mappings.
      operationId: list_roles_admin_roles_get
      security:
      - HTTPBearer: []
      parameters:
      - name: admin_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Admin Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response List Roles Admin Roles Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/roles/{role}:
    get:
      summary: Get Role
      operationId: get_role_admin_roles__role__get
      security:
      - HTTPBearer: []
      parameters:
      - name: role
        in: path
        required: true
        schema:
          type: string
          title: Role
      - name: admin_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Admin Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Role Admin Roles  Role  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      summary: Assign Role
      description: 'Assign an agent to a role.  Triggers registry reload for this
        role.


        This is the only action needed to swap the active agent for a role.

        Existing conversation history is preserved because thread_id is keyed on the
        role.'
      operationId: assign_role_admin_roles__role__put
      security:
      - HTTPBearer: []
      parameters:
      - name: role
        in: path
        required: true
        schema:
          type: string
          title: Role
      - name: admin_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Admin Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleAssign'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Assign Role Admin Roles  Role  Put
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      summary: Unmap Role
      description: Remove a role mapping (does not delete the agent).
      operationId: unmap_role_admin_roles__role__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: role
        in: path
        required: true
        schema:
          type: string
          title: Role
      - name: admin_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Admin Session
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/snippets:
    get:
      summary: List Snippets
      operationId: list_snippets_admin_snippets_get
      security:
      - HTTPBearer: []
      parameters:
      - name: admin_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Admin Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response List Snippets Admin Snippets Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      summary: Create Snippet
      operationId: create_snippet_admin_snippets_post
      security:
      - HTTPBearer: []
      parameters:
      - name: admin_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Admin Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SnippetCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Create Snippet Admin Snippets Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/snippets/{key}:
    put:
      summary: Update Snippet
      description: Update a snippet.  Triggers registry reload for all agents that
        reference it.
      operationId: update_snippet_admin_snippets__key__put
      security:
      - HTTPBearer: []
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
          title: Key
      - name: admin_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Admin Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SnippetUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Update Snippet Admin Snippets  Key  Put
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      summary: Delete Snippet
      description: Delete a snippet.  Fails if any agent references it.
      operationId: delete_snippet_admin_snippets__key__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
          title: Key
      - name: admin_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Admin Session
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      summary: Get Snippet
      operationId: get_snippet_admin_snippets__key__get
      security:
      - HTTPBearer: []
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
          title: Key
      - name: admin_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Admin Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Snippet Admin Snippets  Key  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/observability:
    get:
      summary: Get Observability
      description: Returns the LangFuse dashboard URL for monitoring.
      operationId: get_observability_admin_observability_get
      security:
      - HTTPBearer: []
      parameters:
      - name: admin_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Admin Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Observability Admin Observability Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AgentCreate:
      properties:
        slug:
          type: string
          minLength: 1
          title: Slug
        name:
          type: string
          minLength: 1
          title: Name
        productSlug:
          type: string
          minLength: 1
          title: Productslug
        provider:
          type: string
          pattern: ^(anthropic)$
          title: Provider
          default: anthropic
        model:
          type: string
          minLength: 1
          title: Model
        systemPrompt:
          type: string
          minLength: 1
          title: Systemprompt
        temperature:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Temperature
          default: 0.3
        maxTurns:
          type: integer
          maximum: 100.0
          minimum: 1.0
          title: Maxturns
          default: 25
        tools:
          anyOf:
          - items:
              additionalProperties: true
              type: object
            type: array
          - type: 'null'
          title: Tools
      type: object
      required:
      - slug
      - name
      - productSlug
      - model
      - systemPrompt
      title: AgentCreate
    AgentUpdate:
      properties:
        slug:
          anyOf:
          - type: string
          - type: 'null'
          title: Slug
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        productSlug:
          anyOf:
          - type: string
          - type: 'null'
          title: Productslug
        provider:
          anyOf:
          - type: string
          - type: 'null'
          title: Provider
        model:
          anyOf:
          - type: string
          - type: 'null'
          title: Model
        systemPrompt:
          anyOf:
          - type: string
          - type: 'null'
          title: Systemprompt
        temperature:
          anyOf:
          - type: number
            maximum: 1.0
            minimum: 0.0
          - type: 'null'
          title: Temperature
        maxTurns:
          anyOf:
          - type: integer
            maximum: 100.0
            minimum: 1.0
          - type: 'null'
          title: Maxturns
        tools:
          anyOf:
          - items:
              additionalProperties: true
              type: object
            type: array
          - type: 'null'
          title: Tools
      type: object
      title: AgentUpdate
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    RoleAssign:
      properties:
        agentId:
          type: string
          minLength: 1
          title: Agentid
      type: object
      required:
      - agentId
      title: RoleAssign
    SnippetCreate:
      properties:
        key:
          type: string
          minLength: 1
          pattern: ^[a-z0-9_-]+$
          title: Key
        content:
          type: string
          minLength: 1
          title: Content
      type: object
      required:
      - key
      - content
      title: SnippetCreate
    SnippetUpdate:
      properties:
        content:
          type: string
          minLength: 1
          title: Content
      type: object
      required:
      - content
      title: SnippetUpdate
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer